Trivial freebsd fix.
git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/classlib/trunk@692043 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/luni/src/main/native/luni/unix/freebsd/OSResourcesMonitor.c
@@ -37,7 +37,7 @@ JNIEXPORT jboolean JNICALL Java_org_apache_harmony_luni_platform_OSResourcesMoni
 {
   unsigned long free;
   unsigned long total;
-  int len = sizeof(free);
+  size_t len = sizeof(free);
   
   if (sysctlbyname("vm.stats.vm.v_free_count", &free, &len, NULL, 0) == -1) {
     return FALSE;
